Move mockRoute() from util.js to http.js
